+++ changelog.md
@@ -3,6 +3,7 @@
 ## Version 1.5.0 (Future)
 - Add search filter to entity report
 - Throw different exception in link builder when running in console
+- Don't parse oembed value if it's null in oembed field type
 
 ## Version 1.4.1 (2017-03-14)
 - Remove debug